+++ server/objects/cursor.c
@@ -156,11 +156,11 @@ tw_cursor_unset_wrap(struct tw_cursor *cursor)
 }
 
 void
-tw_cursor_move(struct tw_cursor *cursor, int32_t dx, int32_t dy)
+tw_cursor_move(struct tw_cursor *cursor, float dx, float dy)
 {
 	struct tw_cursor_constrain *constrain;
-	int32_t nx = cursor->x + dx;
-	int32_t ny = cursor->y + dy;
+	float nx = cursor->x + dx;
+	float ny = cursor->y + dy;
 	bool inbound = wl_list_length(&cursor->constrains) ? false : true;
 
 	wl_list_for_each(constrain, &cursor->constrains, link) {
@@ -186,7 +186,7 @@ set_pos:
 }
 
 void
-tw_cursor_move_with_wrap(struct tw_cursor *cursor, int32_t dx, int32_t dy,
+tw_cursor_move_with_wrap(struct tw_cursor *cursor, float dx, float dy,
                          int32_t x, int32_t y, uint32_t width, uint32_t height)
 {
 	pixman_box32_t *e = pixman_region32_extents(&cursor->curr_wrap.region);
@@ -197,7 +197,7 @@ tw_cursor_move_with_wrap(struct tw_cursor *cursor, int32_t dx, int32_t dy,
 }
 
 void
-tw_cursor_set_pos(struct tw_cursor *cursor, int32_t nx, int32_t ny)
+tw_cursor_set_pos(struct tw_cursor *cursor, float nx, float ny)
 {
 	tw_cursor_move(cursor, nx - cursor->x, ny - cursor->y);
 }
